@ankane 2024-03-29
Thanks @Gleonett, this is great! The`Parse` change looks good. For `String`, I'd like to avoid `unsafe` and keep the structure of the current code. I'm seeing similar performance with the following diff:

```diff
 // String returns a string representation of the vector.
 func (v Vector) String() string {
-	var buf strings.Builder
-	buf.WriteString("[")
+	buf := make([]byte, 0, 2+10*len(v.vec))
+	buf = append(buf, '[')
 
 	for i := 0; i < len(v.vec); i++ {
 		if i > 0 {
-			buf.WriteString(",")
+			buf = append(buf, ',')
 		}
-		buf.WriteString(strconv.FormatFloat(float64(v.vec[i]), 'f', -1, 32))
+		buf = strconv.AppendFloat(buf, float64(v.vec[i]), 'f', -1, 32)
 	}
 
-	buf.WriteString("]")
-	return buf.String()
+	buf = append(buf, ']')
+	return string(buf)
 }
```