@Gleonett 2024-03-29
Thanks @ankane for the prompt response. Is 9 bytes per float an empirical figure? Just for the sake of interest.