@ankane 2024-03-29
It looks like 15 bytes per float [would be better](https://github.com/pgvector/pgvector/blob/v0.6.2/src/vector.c#L301) (`FLOAT_SHORTEST_DECIMAL_LEN` is defined as 16). This reduces it to only two allocations.