@Gleonett 2024-03-29
Good, fixed that.

#### P.S.
> Edit: Also, even better performance could be achieved by using the binary format with pgx (with pgtype and pgio). I did some work on this a while ago and plan to revisit it at some point.

I am really looking forward to it!